requirements.txt
argparse
termcolor
python-libnmap